requirements.txt
Flask
mysql-connector-python
bcrypt
pyotp
python-dotenv
psycopg2-binary
